refactor: move top vaults exporter into package
